Next part of changes

+++ src/common/fragment_number.rs
@@ -1,4 +1,4 @@
 #[derive(PartialOrd, PartialEq, Ord, Eq)]
 pub struct FragmentNumber_t {
-    pub vendorId: u64
+    pub vendorId: u32
 }
